Split UI state into dedicated context
