+++ bsky-client/src/modules/profile/ProfileViewerPane.jsx
@@ -1,5 +1,6 @@
 import { Suspense, lazy, useCallback, useState } from 'react'
-import { useAppState, useAppDispatch } from '../../context/AppContext'
+import { useAppDispatch } from '../../context/AppContext'
+import { useUIState } from '../../context/UIContext.jsx'
 import ProfileMetaSkeleton from './ProfileMetaSkeleton.jsx'
 import BskyDetailPane from '../layout/BskyDetailPane.jsx'
 
@@ -9,7 +10,7 @@ const ProfileViewLazy = lazy(async () => {
 })
 
 export default function ProfileViewerPane ({ registerLayoutHeader, renderHeaderInLayout = false }) {
-  const { profileViewer, notificationsSettingsOpen } = useAppState()
+  const { profileViewer, notificationsSettingsOpen } = useUIState()
   const dispatch = useAppDispatch()
   const actor = profileViewer?.actor || ''
   const [headline, setHeadline] = useState({ name: '', handle: '', did: '' })
